Better handling&logging of exception in the validator
Since the validator is implemented as a generator, exceptions in the
online version of the validator were not catched anywhere, and with
gunicorn (in prod), they were not logged nor reported.
A custom version of the stream_with_context generator was implemented
that catch exceptions and send them to flask's logger.

Exceptions in the cron task version of the validator would simply crash
the script. It's fixed now to catch them and print them on stdout.

The validator is not really supposed to raise exceptions; it's supposed to
handle (most of) them and print the corresponding message; but since it's
doing HTTP requests and other complex stuff, a lot could happen and there
are no guarantees all exceptions will be handled.